feat(LT-4): Phase 3 Screen 1 — app-shell status bar live tick
C++ host:
- BridgeDispatcher.EmitStatsTick(fps, emitters, particles, instances).
- HostWindow starts a 4 Hz SetTimer in WM_CREATE; the WM_TIMER handler
  reads FPSMeasurer + Engine counters and emits the stats/tick event.
- FPSMeasurer ported from src/main.cpp (32-frame ring buffer);
  measure() called from RenderD3D9.

React:
- StatusBar component subscribes to stats/tick and renders 4 columns
  (FPS · Emitters · Particles · Instances). Dashes when no host
  (browser mode against MockBridge).
- App.tsx swaps the inline footer for <StatusBar bridge=…/>.

Tests:
- app-shell.spec.ts: 2 specs — StatusBar DOM presence + first stats/tick
  delivery within 1.5s. Bumps native suite to 12 passing.

Design locked in tasks/lt4_design_parking_lot.md Screen 1.

+++ src/host/BridgeDispatcher.h
@@ -85,11 +85,9 @@ public:
     std::string DispatchSync(const std::string& jsonRequest);
 
     // Convenience emitters (host-driven). EmitStatsTick is the 4 Hz status
-    // bar push; EmitEngineStateChanged is the post-setter broadcast. Both
-    // are no-ops in Task 1.3 — listed here so the next slice can wire them
-    // without re-touching the public surface.
+    // bar push; EmitEngineStateChanged is the post-setter broadcast.
     void EmitEngineStateChanged();
-    void EmitStatsTick(int fps, int emitters, int particles, int instances);
+    void EmitStatsTick(float fps, int emitters, int particles, int instances);
 
     // Emits an `accelerator/pressed` event to React with the matched combo
     // string (e.g. "Ctrl+S"). Called by HostWindow's AcceleratorKeyPressed
